backend/etl/requirements.txt
pandas>=2.1.0
numpy>=1.27.0
python-dotenv>=1.0.0
pandas-gbq>=0.26.1
google-cloud-bigquery>=3.10.0
google-cloud-bigquery-storage>=2.20
pyarrow>=12.0.0
db-dtypes>=1.2.0
requests>=2.31.0
yfinance>=0.2.28
python-dateutil>=2.9.0
pandas-market-calendars>=4.4.0